refactor(cli): split flags.ts and input.ts into internal modules
Split flags.ts (909→63 LOC) into flag-types, flag-builders, presets.
Split input.ts (861→99 LOC) into input-security, input-helpers,
input-parsers, input-normalization. Add eslint-disable to types.ts
(types-only file, splitting hurts discoverability).

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
